% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Correlation matrices}
\usage{
calculate_correlations(data, group1, group2, method)
}
\arguments{
\item{data}{The data matrix, with samples on columns and observations on rows}

\item{group1}{column ids for group 1}

\item{group2}{column ids for group 2}

\item{method}{correlation type (defaults to Spearman correlations.)}
}
\value{
returns a `tibble` with the correlations between all the vartiables for both groups. x and y in data frame are row ids that can be easily mapped to variable names in data set.
}
\description{
This function computes the correlation matrices for two groups given a data set (e.g., an expression matrix.) This function uses the `corrr` package to compute the correlations.
}
