% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{cleanup}
\usage{
cleanup(correlation_df)
}
\arguments{
\item{correlation_df}{A corrrelation data frame as calculated with \code{\link{calculate_correlations}}}
}
\value{
A cleaned up correlation matrix
}
\description{
Function eliminates NAs from matrix and sets absolute value of correlations smaller than user defined correlation threshold to 0.
}
