% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_interactions.R
\name{define_interactions}
\alias{define_interactions}
\title{Define interactions}
\usage{
define_interactions(correlation_df, correlation_thr)
}
\arguments{
\item{correlation_df}{A correlation data frame (see \code{\link{calculate_correlations}})}

\item{correlation_thr}{A threshold for correlations (defaults to 0.75)}
}
\value{
An interaction data frame, where the presence or absence of an edge in the correlation network is given as a 1 or a 0 (respectively)
}
\description{
Given a correlation data frame (as calculated by \code{\link{calculate_correlations}}) and a correlation threshold, defines the set of interactions in a correlation adjacency matrix.
}
