% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_top.R
\name{diff_top}
\alias{diff_top}
\title{diff_top}
\usage{
diff_top(data, group1, group2, correlation_type = NULL, correlation_thr)
}
\arguments{
\item{data}{Data matrix, where rows are observations (eg, genes) and columns are samples.}

\item{group1}{Column ids for group 1.}

\item{group2}{Column ids for group 2.}

\item{correlation_type}{Method to calculate correlation.  Defaults to Spearman correlations.}

\item{correlation_thr}{Correlation threshold for edge assignment.  Defaults to 0.75}
}
\value{
Returns a `tibble` that is sorted by the total edge score.
}
\description{
Wrapper for the diffnet package.
}
