% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_test.R
\name{difference_test}
\alias{difference_test}
\title{Correlation difference test}
\usage{
difference_test(correlation_df, g1_size, g2_size)
}
\arguments{
\item{correlation_df}{A correlation data frame}

\item{g1_size}{Size of group 1}

\item{g2_size}{Size of group 2}
}
\value{
A data frame with the test statistic
}
\description{
Given a correlation data frame (calculated using \code{\link{calculate_correlations}}) calculate Fisher's correlation test.
}
