\name{DiffNet_GHD_Fast}
\alias{DiffNet_GHD_Fast}
\docType{data}
\title{
Fast Implementation for Generalized Hamming Distance Statistic
}
\description{
Provides a very fast implementation for generalized hamming distance statistic.
}
%\usage{data("DiffNet_GHD_Fast")}
\format{
  The format is:
List of 4
 $ name         : chr "DiffNet_GHD_Fast"
 $ address      :Class 'RegisteredNativeSymbol' <externalptr> 
 $ dll          :List of 5
  ..$ name         : chr "DiffNet"
  ..$ path         : chr "/home/rmall/R/x86_64-pc-linux-gnu-library/3.3/DiffNet/libs/DiffNet.so"
  ..$ dynamicLookup: logi FALSE
  ..$ handle       :Class 'DLLHandle' <externalptr> 
  ..$ info         :Class 'DLLInfoReference' <externalptr> 
  ..- attr(*, "class")= chr "DLLInfo"
 $ numParameters: int 2
 - attr(*, "class")= chr [1:2] "CallRoutine" "NativeSymbolInfo"
}
\author{
Raghvendra Mall <rmall@hbku.edu.qa>
}


\seealso{
\code{\link{DiffNet_MU_Fast}}, \code{\link{DiffNet_STD_Fast}}
}
