\name{DiffNet_STD_Fast}
\alias{DiffNet_STD_Fast}
\docType{data}
\title{
Fast Implementation of Second Order Moment of Generalized Hamming Distance Statistic
}
\description{
Provides a very fast implementation for obtaining the asymptotic standard deviation for GHD statistic
}
%\usage{data("DiffNet_STD_Fast")}
\format{
  The format is:
List of 4
 $ name         : chr "DiffNet_STD_Fast"
 $ address      :Class 'RegisteredNativeSymbol' <externalptr> 
 $ dll          :List of 5
  ..$ name         : chr "DiffNet"
  ..$ path         : chr "/home/rmall/R/x86_64-pc-linux-gnu-library/3.3/DiffNet/libs/DiffNet.so"
  ..$ dynamicLookup: logi FALSE
  ..$ handle       :Class 'DLLHandle' <externalptr> 
  ..$ info         :Class 'DLLInfoReference' <externalptr> 
  ..- attr(*, "class")= chr "DLLInfo"
 $ numParameters: int 2
 - attr(*, "class")= chr [1:2] "CallRoutine" "NativeSymbolInfo"
}
\author{
Raghvendra Mall <rmall@hbku.edu.qa>
}

\seealso{
\code{\link{GHD_Fast}}, \code{\link{MU_Fast}}
}
