% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_score.R
\name{edge_score}
\alias{edge_score}
\title{Scoring edges}
\usage{
edge_score(scores_df, test_df)
}
\arguments{
\item{scores_df}{A scores data frame as calculated by the \code{\link{score_changes}} function.}

\item{test_df}{A difference test data frame calculated by the \code{\link{difference_test}} function.}
}
\value{
A `tibble`, ranked by total score of the edges
}
\description{
Calculate a total score for the differences in topology.
}
