% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_degree_w_c}
\alias{get_degree_w_c}
\title{get_degree_c}
\usage{
get_degree_w_c(x, W)
}
\arguments{
\item{x}{Covariate vector}

\item{W}{Matrix of [y1+y2, y1+y3, .., y1+yK] to get degree of y1}
}
\description{
This function returns a sum statistic d for vector y tested with all other variables in matrix Y against covariate x
}
