% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_score_c}
\alias{get_score_c}
\title{get_score_c}
\usage{
get_score_c(x, y1, y2)
}
\arguments{
\item{x}{Covariate vector}

\item{y1}{variable 1}

\item{y2}{variable 2}
}
\description{
This function returns the score statistic between y1 and y2 against one-dimensional covariate x
The returned score statistic is not adjusted for sample size
}
