% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_score_W_c}
\alias{get_score_W_c}
\title{get_score_W_c}
\usage{
get_score_W_c(x, W)
}
\arguments{
\item{x}{Covariate vector}

\item{W}{Sums for each pairs of variables}
}
\description{
This function returns a vector of score statistics from one dimensional covariate x and coexpression matrix W
W should be in the form of (y1+y2, y1+y3, ... y1+yK)
The function returns K-1 scores for each pair
}
