\name{GHD_Fast}
\alias{GHD_Fast}
\title{
Fast Implementation for Generalized Hamming Distance Statistic
}
\description{
Provides a very fast implementation for generalized hamming distance statistic.
}
\usage{
GHD_Fast(A, B)
}
\arguments{
  \item{A}{
   Topological matrix obtained from graph g_A.
}
  \item{B}{
   Topological matrix obtained from graph g_B.
}
}
\value{
  Returns the Generalized Hamming Distance between topological matrices of graphs g_A and g_B.
}
\author{
Raghvendra Mall <rmall@hbku.edu.qa>
}


\seealso{
\code{\link{MU_Fast}}, \code{\link{STD_Fast}}
}
