% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generation.R
\name{graph_generation}
\alias{graph_generation}
\title{dummy graph generation}
\usage{
graph_generation(n.nodes = 10, prob.connection = 0.5)
}
\arguments{
\item{n.nodes}{number of nodes}

\item{prob.connection}{node connection probability (default=0.5)}
}
\value{
igraph object
}
\description{
dummy graph generation
}
\details{
Generate a random graph
}
\examples{
graph = graph_generation(n.nodes = 10, prob.connection = 0.5)
initial_score = c(rep(0,5),0.2, 0.3, 0, 0, 0.5)
names(initial_score) = igraph::V(graph)

}
