% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{mvrnormArma}
\usage{
mvrnormArma(n, mu, Sigma)
}
\arguments{
\item{n}{Sample size}

\item{mu}{Mean vector}

\item{Sigma}{Variance-covariance matrix}
}
\description{
This function returns n samples of multivariate normal distribution with mean mu and variance Sigma.
}
