% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_changes.R
\name{score_changes}
\alias{score_changes}
\title{Score topology changes}
\usage{
score_changes(interaction_df)
}
\arguments{
\item{interaction_df}{Interaction data frame}
}
\value{
A score data frame for each edge in the network
}
\description{
Given an interaction data frame (calculated with \code{\link(define_interactions}), provides an edge score and an edge class given the correlation value and the edge presence call.
}
