% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_c}
\alias{simulate_c}
\title{simulate_c}
\usage{
simulate_c(x, Sigma, use_Hhat = TRUE)
}
\arguments{
\item{x}{Covariate vector}

\item{Sigma}{true covariance matrix of matrix Y}

\item{use_Hhat}{Boolean argument whether to use true Sigma or estimate Sigma}
}
\description{
Given fixed covariate and covariance matrix, this function simulates multivariate normal Y
and returns the score vector for each variable
}
