% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{small_sample_correction_c}
\alias{small_sample_correction_c}
\title{small_sample_correction_c}
\usage{
small_sample_correction_c(x, C, k, p = 2L)
}
\arguments{
\item{x}{Covariate vector}

\item{C}{chi-squared threshold for desired FDR rate}

\item{k}{number of covariates + 1}

\item{p}{number of covariates + 1}
}
\description{
This function returns a sum statistic d for vector y tested with all other variables in matrix Y against covariate x
}
