% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet.R
\docType{package}
\name{diffnet}
\alias{diffnet}
\alias{diffnet-package}
\title{DiffNet}
\description{
Calculating the topological differences in a network based on an omics data set. This approach relies on the calculation of the correlation between the variables in the data set and on Fisher's correlation test to define how edges influence changes in topology.
}
